Remove duplicate API implementation
