--- op-item-db-api/package.json
@@ -1,17 +1,0 @@
-{
-	"name": "op-item-db-api",
-	"type": "module",
-	"scripts": {
-		"dev": "wrangler dev",
-		"deploy": "wrangler deploy --minify",
-		"cf-typegen": "wrangler types --env-interface CloudflareBindings"
-	},
-	"dependencies": {
-		"@supabase/supabase-js": "^2.57.4",
-		"hono": "^4.9.8",
-		"zod": "^4.1.11"
-	},
-	"devDependencies": {
-		"wrangler": "^4.38.0"
-	}
-}
