--- op-item-db-api/src/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "src",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "type": "commonjs",
-  "dependencies": {
-    "@supabase/supabase-js": "2.57.4",
-    "hono": "4.9.8"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "4.20250923.0",
-    "typescript": "5.9.2",
-    "wrangler": "4.38.0"
-  }
-}
