--- op-item-db-api/wrangler.toml
@@ -1,14 +1,0 @@
-name = "op-item-db-api"
-main = "src/index.ts"
-compatibility_date = "2024-11-01"
-
-# Standalone Worker, der /api/* auf deiner Domain bedient:
-routes = [
-  { pattern = "op-item-db.com/api/*", zone_name = "op-item-db.com" }
-]
-
-[vars]
-SUPABASE_URL = "https://pgrtmdxmewsxixzubqpr.supabase.co"
-
-# Falls supabase-js Node-APIs braucht (meist nicht nötig):
-# node_compat = true
